r Extract text.c
