+++ gemtext_to_html/Makefile
@@ -17,22 +17,26 @@ build/gemtext_to_html_test: src/arena.h
 build/gemtext_to_html_test: src/astring.c
 build/gemtext_to_html_test: src/astring.h
 build/gemtext_to_html_test: src/testing.h
+build/gemtext_to_html_test: src/text.c
 build/gemtext_to_html_test: src/text.h
 build/gemtext_to_html_test: | build
 	cc -g -Og -o "$@" \
 		src/gemtext_to_html_test.c \
 		src/gemtext_to_html.c \
 		src/arena.c \
-		src/astring.c
+		src/astring.c \
+		src/text.c
 
 build/text_test: src/text_test.c
+build/text_test: src/text.c
 build/text_test: src/text.h
 build/text_test: src/arena.c
 build/text_test: src/arena.h
 build/text_test: | build
 	cc -g -Og -o "$@" \
 		src/text_test.c \
-		src/arena.c
+		src/arena.c \
+		src/text.c
 
 .PHONY: mem_check
 mem_check:
