+++ gemtext_to_html/src/text.h
@@ -2,28 +2,12 @@
 #define TEXT_H
 
 #include <stdbool.h>
-#include <stdio.h>
-#include <string.h>
 
-bool starts_with(char *text, char *prefix) {
-    if (text && prefix) {
-        char *result = strstr(text, prefix);
-        if (result && result == text) {
-            return true;
-        }
-    }
-    return false;
-}
+#define SPACES " \f\n\r\t\v"
 
-static const char *spaces = " \f\n\r\t\v";
-
-void skip_spaces(char **s) {
-    *s += strspn(*s, spaces);
-}
-
-void skip_non_spaces(char **s) {
-    *s += strcspn(*s, spaces);
-}
+bool starts_with(char *text, char *prefix);
+void skip_spaces(char **s);
+void skip_non_spaces(char **s);
 
 #define print(fmt, ...) (printf(fmt"\n", __VA_ARGS__))
